# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: SIRCCD API
  description: |
    API REST para el Sistema Inteligente Urbano para Reporte y Priorización de Daños Viales.
    
    Permite gestionar reportes ciudadanos, clasificación automática mediante IA, 
    deduplicación, priorización multicriterio y gestión operativa municipal.
  version: 1.0.0
  contact:
    name: Equipo SIRCCD
    email: soporte@sirccd.gob.sv
servers:
  - url: http://localhost:8000/api/v1
    description: Desarrollo local
  - url: https://api.sirccd.gob.sv/api/v1
    description: Producción
...
paths:
...
  /auth/login:
    post:
      tags:
        - auth
      summary: Iniciar sesión
      description: Autenticación de usuario (ciudadano, supervisor, administrador)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: usuario@example.com
                password:
                  type: string
                  format: password
                  example: Password123!
      responses:
        '200':
          description: Autenticación exitosa
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                  refresh_token:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
                  token_type:
                    type: string
                    example: Bearer
                  expires_in:
                    type: integer
                    example: 3600
                  user:
                    $ref: '#/components/schemas/Usuario'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Usuario:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        nombre:
          type: string
        rol:
          type: string
          enum: [ciudadano, administrador, supervisor]
        activo:
          type: boolean
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
          nullable: true
  responses:
...
    Unauthorized:
      description: No autorizado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: unauthorized
            message: Token de autenticación inválido o expirado
...
    ValidationError:
      description: Error de validación
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/Error'
              - type: object
                properties:
                  details:
                    type: array
                    items:
                      type: object
                      properties:
                        field:
                          type: string
                        message:
                          type: string
          example:
            error: validation_error
            message: Los datos proporcionados no son válidos
            details:
              - field: email
                message: Formato de email inválido
              - field: password
                message: La contraseña debe tener al menos 8 caracteres
security:
  - bearerAuth: []